Remove unused snap code
