--- Assets/Editor/tests/model/core/SnapGridTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e429e409a14911045b384915ce5a7501
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
