+++ Assets/Scripts/model/controller/PaletteController.cs
@@ -408,7 +408,7 @@ namespace com.google.apps.peltzer.client.model.controller
                     }
                 }
 
-                if (Features.useContinuousSnapDetection && controller.IsTriggerHalfPressed())
+                if (controller.IsTriggerHalfPressed())
                 {
                     if (PaletteControllerActionHandler != null)
                     {
@@ -420,7 +420,7 @@ namespace com.google.apps.peltzer.client.model.controller
                     }
                 }
 
-                if (Features.useContinuousSnapDetection && controller.WasTriggerJustReleasedFromHalfPress())
+                if (controller.WasTriggerJustReleasedFromHalfPress())
                 {
                     if (PaletteControllerActionHandler != null)
                     {
