--- Assets/Scripts/model/render/UXEffects/FaceSnapEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 090a64834cc69194e9a3130c1bb2a107
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
